add chart thong ke thu nhap theo thang trong nam

+++ T1809E_Project_Sem3/Models/ModelsChart.cs
@@ -11,7 +11,7 @@ namespace T1809E_Project_Sem3.Models
         [DataContract]
         public class DataPoint
         {
-            public DataPoint(double x, double y)
+            public DataPoint(int x, double y)
             {
                 this.X = x;
                 this.Y = y;
@@ -19,11 +19,11 @@ namespace T1809E_Project_Sem3.Models
 
             //Explicitly setting the name to be used while serializing to JSON.
             [DataMember(Name = "x")]
-            public Nullable<double> X = null;
+            public Nullable<int> X = null;
 
             //Explicitly setting the name to be used while serializing to JSON.
             [DataMember(Name = "y")]
-            public Nullable<double> Y = null;
+            public Nullable<Double> Y = null;
         }
 	}
 }
